Rename piperlib to libpiper.
